requirements.txt
aiogram
git+https://github.com/krypton-byte/anonclient
python-dotenv==0.20.0